Fixed SpeechColor and added SpeechColorOverride property. Fix for Enhanced Client colored speech. Fixed MulDiv (undefined keyword) and ACTION (getting wrong value when it was -1). Removed stlport.zip, since we're using C++ 11 by default we don't need it anymore. Renamed NCHAR data type to nchar, some other refactoring/tidying and minor fixes. Added HUE_SAY_DEF enum with player default yellow speech hue. Added comments explaining some code.
Details:
Fixed: speech color for NPCs on Enhanced Clients being always "yellow". Achieved this by forcing (only for EC) system (0) talkmode instead of say (3).
	This workaround works only if the speech packet sent by Sphere is Unicode (instead of ASCII), so you'll need to set SAY_DEF_UNICODE in sphere_msgs.scp to 1.
Added: SpeechColorOverride property (players only, R/W) to override client speech hue. This will not affect Words of Power hue, you can use spell triggers for this
	(WOP hue on Enhanced Client will still be set by the client, regardless of the hue sent by the server).
	SpeechColorOverride = 52 (or 0x34), which is the default yellow hue for player speech, disables the override.
	Remember: SpeechColor property is read-only for players and R/W for NPCs. For players, it contains the last speech hue sent by the client. For NPCs, its value
		will be used to set their speech hue.
Fixed: SpeechColor not overriding the speech hue for NPCs. SpeechColor = 946 (or 0x382), which is the default gray hue for NPC speech, disables the override.
Changed: If a hue is not specified, SAY and MSG will use SpeechColor or, if set, SpeechColorOverride.

+++ src/common/CResourceBase.cpp
@@ -288,8 +288,8 @@ lpctstr CResourceBase::ResourceGetName( RESOURCE_ID_BASE rid ) const
 
 	tchar * pszTmp = Str_GetTemp();
 	ASSERT(pszTmp);
-	if ( ! rid.IsValidUID() )
-		sprintf( pszTmp, "%" PRIu32, rid.GetPrivateUID() );
+	if ( !rid.IsValidUID() )
+		sprintf( pszTmp, "%" PRId32, (int)rid.GetPrivateUID() );	// casting to int: it's not an UID, so it's not unsigned
 	else
 		sprintf( pszTmp, "0%" PRIx32, rid.GetResIndex() );
 	return pszTmp;
